% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{areNatIntsOrNAs}
\alias{areNatIntsOrNAs}
\title{Returns true iff all the xs are naturals (>= 0) or NAs}
\usage{
areNatIntsOrNAs(xs)
}
\arguments{
\item{xs}{vector to check}
}
\value{
true or false
}
\description{
Returns true iff all the xs are naturals (>= 0) or NAs
}
