% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{areNegInts}
\alias{areNegInts}
\title{Returns true iff all the xs are negative}
\usage{
areNegInts(xs)
}
\arguments{
\item{xs}{vector to check}
}
\value{
true or false
}
\description{
Returns true iff all the xs are negative
}
