% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{areNegIntsOrNAs}
\alias{areNegIntsOrNAs}
\title{Returns true iff all the xs are negative or NAs}
\usage{
areNegIntsOrNAs(xs)
}
\arguments{
\item{xs}{vector to check}
}
\value{
true or false
}
\description{
Returns true iff all the xs are negative or NAs
}
