% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{areNonNegDoubles}
\alias{areNonNegDoubles}
\title{Returns true iff all the xs are non-negative}
\usage{
areNonNegDoubles(xs)
}
\arguments{
\item{xs}{vector to check}
}
\value{
true or false
}
\description{
Returns true iff all the xs are non-negative
}
