% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arePosIntsOrNAs}
\alias{arePosIntsOrNAs}
\title{Returns true iff all the xs are positive or NAs}
\usage{
arePosIntsOrNAs(xs)
}
\arguments{
\item{xs}{vector to check}
}
\value{
true or false
}
\description{
Returns true iff all the xs are positive or NAs
}
