% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_factors.R
\name{calc_factors}
\alias{calc_factors}
\title{Calculate health factors from the 2022 county health rankings model}
\usage{
calc_factors(chr, measures, z_state = F)
}
\arguments{
\item{chr}{a county health rankings analytic data release (coerced to data.table if not one already)}

\item{measures}{a list of the measures}

\item{z_state}{if TRUE then the measures are scaled (z-scored) at the state level (used for ranking counties within a state)
if FALSE (default) measures are scaled relative to national mean and SD}
}
\value{
\code{chr} as a data.table with additional columns reflecting the factors specified in measures (see names(measures) for names of health factors)
}
\description{
Given the columns and weights specified in measures, calculate health factors used in
county health rankings. Currently get_data and get_measures only support 2022
data. The results of get_measures can be modified to alter the variables and weights
used in calculating the scores. Note the scores themselves, not the rankings, are returned.
To use the scores to create state level rankings, set z_state to TRUE.
}
