% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chEither}
\alias{chEither}
\title{Either ch(eck) where the left and right types are expressed by checks
cl and cr.}
\usage{
chEither(cl, cr, x, asPred = FALSE)
}
\description{
Either ch(eck) where the left and right types are expressed by checks
cl and cr.
}
