% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chEmptyString}
\alias{chEmptyString}
\title{\code{s == ""} ch(eck) for String, deliberately not chReg-ed}
\usage{
chEmptyString(x, asPred = FALSE)
}
\description{
\code{s == ""} ch(eck) for String, deliberately not chReg-ed
}
