% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chNonBlank}
\alias{chNonBlank}
\title{Non blank-ness ch(eck) for String, deliberately not chReg-ed}
\usage{
chNonBlank(x, asPred = FALSE)
}
\description{
Non blank-ness ch(eck) for String, deliberately not chReg-ed
}
