% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chNonEmptyString}
\alias{chNonEmptyString}
\title{\code{s != ""} ch(eck) for String, deliberately not chReg-ed}
\usage{
chNonEmptyString(x, asPred = FALSE)
}
\description{
\code{s != ""} ch(eck) for String, deliberately not chReg-ed
}
