% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr-package.R
\docType{package}
\name{chr-package}
\alias{chr}
\alias{chr-package}
\title{chr: Simple String Manipulation}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Clean, wrangle, and parse character [string] vectors
using base exclusively base R functions.
}
\author{
\strong{Maintainer}: Michael Wayne Kearney \email{kearneymw@missouri.edu}

}
\keyword{internal}
