% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{chr_count}
\alias{chr_count}
\title{Count matches from strings}
\usage{
chr_count(x, pat, ignore.case = FALSE, invert = FALSE, ...)
}
\arguments{
\item{x}{Character vector}

\item{pat}{Pattern (regex) to extract from text.}

\item{ignore.case}{Logical indicating whether to ignore capitalization.
Defaults to false.}

\item{invert}{Logical indicating whether to extract matching portion
(default) or, if this value is true, non-matching portions of text.}

\item{...}{Other named arguments passed to \code{\link{gregexpr}}.}
}
\value{
Vector of matches extracted from input text.
}
\description{
Count all pattern matches in character vector.
}
