% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{chr_detect}
\alias{chr_detect}
\title{Detect matches in strings}
\usage{
chr_detect(x, pat, ignore.case = FALSE, ...)
}
\arguments{
\item{x}{Character vector}

\item{pat}{Pattern (regex) to detect from text.}

\item{ignore.case}{Logical indicating whether to ignore capitalization.
Defaults to false.}

\item{...}{Other named arguments passed to \code{\link{grepl}}.}
}
\value{
Logical vector indicating whether each element matched the supplied pattern.
}
\description{
Detect matching pattern in character vector.
}
