% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{chr_extract_hashtags}
\alias{chr_extract_hashtags}
\title{Extract hashtags from text}
\usage{
chr_extract_hashtags(x, collapse = NULL)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector of hashtags.
}
\description{
Extracts all hashtags from character vector.
}
