% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{chr_extract_links}
\alias{chr_extract_links}
\title{Extract URL links from text}
\usage{
chr_extract_links(x, collapse = NULL)
}
\arguments{
\item{x}{Character vector.}

\item{collapse}{Text inserted between extracted matches. If non-null (the
default) a vector of matches is returned for each inputted string.}
}
\value{
Vector of URLs.
}
\description{
Extracts all hyper-links from character vector.
}
