% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{chr_extract_mentions}
\alias{chr_extract_mentions}
\title{Extract [at] mentions from text}
\usage{
chr_extract_mentions(x, collapse = NULL)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector of screen names.
}
\description{
Extracts all [at] mentions from character vector.
}
