% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{chr_extract_words}
\alias{chr_extract_words}
\title{Extract words from text}
\usage{
chr_extract_words(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector of words.
}
\description{
Extracts all words from character vector.
}
