% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{chr_ngram_char}
\alias{chr_ngram_char}
\title{Character n-grams}
\usage{
chr_ngram_char(x, n = 3, lower = FALSE, space = FALSE, punct = FALSE)
}
\arguments{
\item{x}{Character vector}

\item{n}{Number of characters to return per ngram}

\item{lower}{Logical indicating whether to lower case all text, defaults to
false.}

\item{space}{Logical indicating whether to strip space, defaults to false.}

\item{punct}{Logical indicating whether to strip punctation, defaults to
false.}
}
\value{
List of length equal to input length consisting of ngram vectors.
}
\description{
Returns n-grams at the character level
}
\details{
Thanks to ChrisMuir \(https://github.com/mkearney/chr/issues/1)
}
\author{
ChrisMuir
}
