% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{chr_remove_hashtags}
\alias{chr_remove_hashtags}
\title{Remove hashtags from text}
\usage{
chr_remove_hashtags(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector without hashtags.
}
\description{
Removes all hashtags from character vector.
}
