% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{chr_remove_linebreaks}
\alias{chr_remove_linebreaks}
\title{Remove line breaks from text}
\usage{
chr_remove_linebreaks(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector without line breaks.
}
\description{
Removes all line breaks from character vector.
}
