% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{chr_remove_links}
\alias{chr_remove_links}
\title{Remove URL links from text}
\usage{
chr_remove_links(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector without URLs.
}
\description{
Removes all hyper-links from character vector.
}
