% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{chr_remove_mentions}
\alias{chr_remove_mentions}
\title{Remove [at] mentions from text}
\usage{
chr_remove_mentions(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector without screen names.
}
\description{
Removes all [at] mentions from character vector.
}
