% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{chr_remove_tabs}
\alias{chr_remove_tabs}
\title{Remove tabs from text}
\usage{
chr_remove_tabs(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector without tabs.
}
\description{
Removes all tabs from character vector.
}
