% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{chr_remove_ws}
\alias{chr_remove_ws}
\title{Remove extra spaces from text}
\usage{
chr_remove_ws(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Vector without extra spaces.
}
\description{
Removes double+ spaces and trims white space from string ends.
}
