% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{chr_replace}
\alias{chr_replace}
\title{Replace text pattern with string}
\usage{
chr_replace(x, m, r, ignore.case = FALSE, ...)
}
\arguments{
\item{x}{Character vector.}

\item{m}{Matching text or regular expression usd to locate text to be
replaced.}

\item{r}{Replacement text, the length of which must be either one (a single
string used as the replacement text) or equal to the length of the supplied
character vector}

\item{ignore.case}{Logical indicating whether to ignore capitalization,
defaults to FALSE.}

\item{...}{Other args passed on to \code{\link{gsub}}.}
}
\value{
Vector without URLs.
}
\description{
Replaces all matching patterns with user-provided string.
}
