% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii.R
\name{chr_replace_nonascii}
\alias{chr_replace_nonascii}
\title{Replace non-ascii with similar ascii characters}
\usage{
chr_replace_nonascii(x)
}
\arguments{
\item{x}{Character vector with non-ascii characters}
}
\value{
ASCII-friendly character vector.
}
\description{
Replace weird UTF values with equivalent(ish) ascii values.
}
