% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chReg}
\alias{chReg}
\title{Registeres the ch(eck) using an optional name (ch(eck) name by default)}
\usage{
chReg(ch, name = NA)
}
\description{
Registeres the ch(eck) using an optional name (ch(eck) name by default)
}
