% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chScalar}
\alias{chScalar}
\title{Scalar \code{is.atomic} & \code{length == 1L} value ch(eck)}
\usage{
chScalar(x, asPred = FALSE)
}
\description{
Scalar \code{is.atomic} & \code{length == 1L} value ch(eck)
}
