% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chsDiff}
\alias{chsDiff}
\title{Returns a vector of ch(eck)s names passed by the first argument, but not
passed by the rest of the args}
\usage{
chsDiff(...)
}
\description{
Returns a vector of ch(eck)s names passed by the first argument, but not
passed by the rest of the args
}
