% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve the 2022 county health rankings analytic dataset}
\usage{
get_data(year = NULL)
}
\arguments{
\item{year}{not currently used as only 2022 variables are returned. Future iterations
may allow for previous versions of the CHR to be calculated}
}
\value{
a data.table containing the CHR data
}
\description{
Gets the 2022 release from https://www.countyhealthrankings.org/ and returns as a data.table
}
