% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Get measures that go into county health rankings}
\usage{
get_measures(year = NULL)
}
\arguments{
\item{year}{not currently used as only 2022 variables are returned. Future iterations
may allow for previous versions of the CHR to be calculated}
}
\value{
a list of health factors. Each health factor is also a list containing
\item{variables} the names of the variables that go in to a particular factor
\item{names} the column names of the variables in the CHR data release
\item{weights} the weight given to each variable in the CHR model
}
\description{
Returns the measures from county health rankings data
that are used to create four 'health factors'. Also returns
the weights for each measure.
}
