% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_case}
\alias{title_case}
\title{Convert string to title case}
\usage{
title_case(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector in title case.
}
\description{
Capitalize character vector using title case
}
