% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{chr_detect}
\alias{chr_detect}
\title{Detect matches in strings}
\usage{
chr_detect(x, pat, ignore.case = FALSE, ...)
}
\arguments{
\item{x}{Character vector}

\item{pat}{Pattern (regex) to detect from text.}

\item{ignore.case}{Logical indicating whether to ignore capitalization.
Defaults to false.}

\item{...}{Other named arguments passed to \code{\link{grepl}} or \code{\link{grep}}
See details for more information.}
}
\value{
Logical vector indicating whether each element matched the supplied pattern.
}
\description{
Detect matching pattern in character vector.
}
\details{
This is a wrapper around the base R functions \code{\link{grepl}} and
  \code{\link{grep}}. By default, logical values are returned (a la grepl). To return
  values, include \code{value = TRUE}. To return positions, include \code{which = TRUE},
  \code{pos = TRUE}, or \code{position = TRUE}.
}
\examples{

## return logical vector
chr_detect(letters, "a|b|c|x|y|z")

## return inverted logical values
chr_detect(letters, "a|b|c|x|y|z", invert = TRUE)

## return matching positions
chr_detect(letters, "a|b|c|x|y|z", which = TRUE)

## return inverted matching positions
chr_detect(letters, "a|b|c|x|y|z", which = TRUE, invert = TRUE)

## return matching values
chr_detect(letters, "a|b|c|x|y|z", value = TRUE)

## return inverted matching values
chr_detect(letters, "a|b|c|x|y|z", value = TRUE, invert = TRUE)
}
