% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{chr_extract}
\alias{chr_extract}
\title{Extract matches from strings}
\usage{
chr_extract(x, pat, ignore.case = FALSE, collapse = NULL, invert = FALSE,
  na = TRUE, ...)
}
\arguments{
\item{x}{Character vector}

\item{pat}{Pattern (regex) to extract from text.}

\item{ignore.case}{Logical indicating whether to ignore capitalization.
Defaults to false.}

\item{collapse}{Text inserted between extracted matches. If non-null (the
default) a vector of matches is returned for each inputted string.}

\item{invert}{Logical indicating whether to extract matching portion
(default) or, if this value is true, non-matching portions of text.}

\item{na}{Logical indicating whether to return NA values for input elements
without matches. Defaults to true.}

\item{...}{Other named arguments passed to \code{\link{gregexpr}}.}
}
\value{
Vector of matches extracted from input text.
}
\description{
Detect and return all matching patterns from character vector.
}
